fix(tests): fixed tests
